mqtt/requirements.txt
paho-mqtt>=1.6.0
pymodbus>=3.0.0
